# systems/if_line/docs/api/story-path-authoring.openapi.yaml
openapi: 3.1.0
info:
  title: IF Line StoryPath Authoring API
  version: 1.0.0
  description: >-
    Replacement authoring and publication contract. This contract is mounted
    directly below /api. Legacy chapter-index authoring routes are not part of
    this API and receive no compatibility adapter.
servers:
  - url: /api
security:
  - cookieAuth: []
...
paths:
  /projects:
    post:
      tags: [Projects]
      operationId: createProject
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/ProjectCreate'}
      responses:
        '201': {$ref: '#/components/responses/Project'}
        '422': {$ref: '#/components/responses/ValidationError'}
    get:
      tags: [Projects]
      operationId: listProjects
      parameters:
        - name: publication_state
          in: query
          schema:
            type: string
            enum: [all, unpublished, published, changes_pending]
            default: all
      responses:
        '200':
          description: Owner project list.
          content:
            application/json:
              schema:
                type: array
                items: {$ref: '#/components/schemas/ProjectRead'}
  /projects/{project_id}:
    parameters:
      - {$ref: '#/components/parameters/ProjectId'}
    get:
      tags: [Projects]
      operationId: getProject
      responses:
        '200': {$ref: '#/components/responses/Project'}
        '404': {$ref: '#/components/responses/NotFound'}
    patch:
      tags: [Projects]
      operationId: updateProject
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/ProjectPatch'}
      responses:
        '200': {$ref: '#/components/responses/Project'}
        '404': {$ref: '#/components/responses/NotFound'}
    delete:
      tags: [Projects]
      operationId: deleteProject
      responses:
        '204': {description: Project deleted or tombstoned.}
        '409': {$ref: '#/components/responses/Conflict'}
...
  /projects/{project_id}/bible-generations:
    parameters:
      - {$ref: '#/components/parameters/ProjectId'}
    post:
      tags: [Bible]
      operationId: generateBible
      parameters:
        - {$ref: '#/components/parameters/IdempotencyKey'}
      requestBody:
        content:
          application/json:
            schema: {$ref: '#/components/schemas/GenerationRequest'}
      responses:
        '202': {$ref: '#/components/responses/TaskAccepted'}
        '409': {$ref: '#/components/responses/Conflict'}
  /projects/{project_id}/bible-revisions:
    parameters:
      - {$ref: '#/components/parameters/ProjectId'}
    get:
      tags: [Bible]
      operationId: listBibleRevisions
      responses:
        '200':
          description: Bible revision history.
          content:
            application/json:
              schema:
                type: array
                items: {$ref: '#/components/schemas/BibleRevision'}
    post:
      tags: [Bible]
      operationId: createBibleRevision
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/BibleRevisionCreate'}
      responses:
        '201':
          description: Unselected ready revision.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/BibleRevision'}
  /projects/{project_id}/bible-head:
    parameters:
      - {$ref: '#/components/parameters/ProjectId'}
    get:
      tags: [Bible]
      operationId: getBibleHead
      responses:
        '200': {$ref: '#/components/responses/RevisionHead'}
        '404': {$ref: '#/components/responses/NotFound'}
    put:
      tags: [Bible]
      operationId: updateBibleHead
      parameters:
        - {$ref: '#/components/parameters/IfMatch'}
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/HeadUpdate'}
      responses:
        '200': {$ref: '#/components/responses/RevisionHead'}
        '409': {$ref: '#/components/responses/Conflict'}
...
  /public/projects/{project_id}:
    parameters:
      - {$ref: '#/components/parameters/ProjectId'}
    get:
      security: []
      tags: [Public]
      operationId: getPublicProject
      responses:
        '200':
          description: Public projection frozen in the active Release.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/PublicProject'}
        '404': {$ref: '#/components/responses/NotFound'}
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: sid
  parameters:
    ProjectId:
      name: project_id
      in: path
      required: true
      schema: {type: integer, minimum: 1}
...
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema: {type: string, minLength: 1, maxLength: 255}
    IfMatch:
      name: If-Match
      in: header
      required: true
      description: Current integer lock version, optionally quoted.
      schema: {type: string, pattern: '^("[1-9][0-9]*"|[1-9][0-9]*)$'}
  responses:
    Project:
      description: Owner project projection.
      content:
        application/json:
          schema: {$ref: '#/components/schemas/ProjectRead'}
    RevisionHead:
      description: Current reviewed revision pointer.
      content:
        application/json:
          schema: {$ref: '#/components/schemas/RevisionHead'}
    TaskAccepted:
      description: Durable asynchronous task accepted.
      content:
        application/json:
          schema: {$ref: '#/components/schemas/TaskAccepted'}
...
    NotFound:
      description: Resource is absent or not visible to this caller.
      content:
        application/json:
          schema: {$ref: '#/components/schemas/Error'}
    Conflict:
      description: Lock, idempotency, provenance, or source fingerprint conflict.
      content:
        application/json:
          schema: {$ref: '#/components/schemas/Error'}
    ValidationError:
      description: Request or publication readiness validation failed.
      content:
        application/json:
          schema: {$ref: '#/components/schemas/Error'}
  schemas:
    UUID:
      type: string
      format: uuid
    ProjectCreate:
      type: object
      required: [title]
      properties:
        title: {type: string, minLength: 1, maxLength: 200}
        characters: {type: array, items: {type: object, additionalProperties: true}}
        story_start: {type: string}
        story_end: {type: string}
        style: {type: string}
        source_work: {type: [string, 'null']}
        pace: {type: [string, 'null']}
        extra_requirements: {type: [string, 'null']}
    ProjectPatch:
      type: object
      minProperties: 1
      properties:
        title: {type: string, minLength: 1, maxLength: 200}
        characters: {type: array, items: {type: object, additionalProperties: true}}
        story_start: {type: string}
        story_end: {type: string}
        style: {type: string}
        source_work: {type: [string, 'null']}
        pace: {type: [string, 'null']}
        extra_requirements: {type: [string, 'null']}
    ProjectRead:
      allOf:
        - {$ref: '#/components/schemas/ProjectCreate'}
        - type: object
          required: [id, authoring, publication]
          properties:
            id: {type: integer}
            root_story_path_id: {$ref: '#/components/schemas/UUID'}
            authoring: {$ref: '#/components/schemas/AuthoringProjection'}
            publication: {$ref: '#/components/schemas/PublicationProjection'}
    AuthoringProjection:
      type: object
      required: [stage, blocking_items, running_task_count, has_unpublished_changes]
      properties:
        stage:
          type: string
          enum: [bible, story_paths, outlines, chapters, scripts, vn_graphs, ready, blocked]
        blocking_items: {type: array, items: {$ref: '#/components/schemas/BlockingItem'}}
        running_task_count: {type: integer, minimum: 0}
        has_unpublished_changes: {type: boolean}
    PublicationProjection:
      type: object
      required: [state, active_release_id, published_at, lock_version]
      properties:
        state: {type: string, enum: [unpublished, published, changes_pending]}
        active_release_id: {oneOf: [{$ref: '#/components/schemas/UUID'}, {type: 'null'}]}
        published_at: {type: [string, 'null'], format: date-time}
        lock_version:
          type: integer
          minimum: 1
          description: Current ProjectPublication version used by unpublish If-Match.
    BlockingItem:
      type: object
      required: [code]
      properties:
        code: {type: string}
        story_path_id: {oneOf: [{$ref: '#/components/schemas/UUID'}, {type: 'null'}]}
        path_chapter_id: {oneOf: [{$ref: '#/components/schemas/UUID'}, {type: 'null'}]}
        detail: {type: [string, 'null']}
...
    GenerationRequest:
      type: object
      properties:
        instructions: {type: [string, 'null'], maxLength: 12000}
        parameters: {type: object, additionalProperties: true}
    BibleRevisionCreate:
      type: object
      required: [content_json]
      properties:
        parent_revision_id: {oneOf: [{$ref: '#/components/schemas/UUID'}, {type: 'null'}]}
        content_json: {type: object, additionalProperties: true}
    BibleRevision:
      type: object
      required: [id, project_id, revision_no, content_hash, content_json, created_at]
      properties:
        id: {$ref: '#/components/schemas/UUID'}
        project_id: {type: integer}
        parent_revision_id: {oneOf: [{$ref: '#/components/schemas/UUID'}, {type: 'null'}]}
        revision_no: {type: integer, minimum: 1}
        source_hash: {type: string}
        content_hash: {type: string}
        content_json: {type: object, additionalProperties: true}
        created_at: {type: string, format: date-time}
    RevisionHead:
      type: object
      required: [revision_id, lock_version]
      properties:
        revision_id: {oneOf: [{$ref: '#/components/schemas/UUID'}, {type: 'null'}]}
        lock_version: {type: integer, minimum: 1}
        updated_at: {type: string, format: date-time}
    HeadUpdate:
      type: object
      required: [revision_id]
      properties:
        revision_id: {$ref: '#/components/schemas/UUID'}
...
    PublicProject:
      type: object
      required: [id, title, release_id, release_version, manifest_hash, published_at]
      properties:
        id: {type: integer}
        title: {type: string}
        summary: {type: string}
        cover_url: {type: [string, 'null']}
        release_id: {$ref: '#/components/schemas/UUID'}
        release_version: {type: integer}
        manifest_hash: {type: string}
        published_at: {type: string, format: date-time}
    TaskAccepted:
      type: object
      required: [task_id, status, events_url, created]
      properties:
        task_id: {$ref: '#/components/schemas/UUID'}
        status: {type: string}
        events_url: {type: string}
        created: {type: boolean}
    Error:
      type: object
      required: [code, message]
      properties:
        code: {type: string}
        message: {type: string}
        details: {type: object, additionalProperties: true}